src/requirements.txt
# requirements.txt

cffi==1.17.0
colorama==0.4.6
colorclass==2.2.2
cryptography==43.0.0
easygui==0.98.3
filetype==1.2.0
lxml==5.3.0
msoffcrypto-tool==5.4.2
olefile==0.47
oletools==0.60.2
pcodedmp==1.2.6
pefile==2023.2.7
pycparser==2.22
pyfiglet==1.0.2
pyparsing==3.1.2
PyPDF2==3.0.1
python-docx==1.1.2
python-magic==0.4.27
python-magic-bin==0.4.14
tabulate==0.9.0
typing_extensions==4.12.2
win_unicode_console==0.5
